Fix inconsistency of condition codes with OPQ
According to the CS:APP book 3rd edition, section 3.6.1, it is stated "For the logical operations, such as XOR, the carry and overflow flags are set to zero". To my understanding, other than not having the carry flag, y86 should have the same behavior on condition codes as x86.

Before, if you do something like an addq that sets the overflow flag, the xorq instruction does not clear it. This commit fixes the inconsistency by first setting OF to zero before going into the switch statement.